Issues Resolved
